Desarrollo del ejercicio 2 y corrección del ejercicio 1 linea 7 por conversion explicita al tipo float

+++ tp1_VillagranEnzoMauricio/src/ar/edu/unju/fi/ejercicio1/Main.java
@@ -4,7 +4,7 @@ public class Main {
 
 	public static void main(String[] args) {
 		int numA=8,numB=5,numC=1,numD=20,numE=9;
-		double totalPromedio = (numA+numB+numC+numD+numE)/5.;
+		float totalPromedio = (numA+numB+numC+numD+numE)/5.0f;
 		System.out.println("El promedio es: "+totalPromedio);
 
 	}
